+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.12.3]
+
+### Added
+
+* `MOC.__init__` and `STMOC.__init__` raise `PermissionError` if user tries to modify order manually
+
+### Fixed
+
+* :bug: position angle limited to PI/2 instead of PI in `MOC.from_elliptical_cone`
+
 ## [0.12.2]
 
 ### Added